Contribute jarmin
This PR contributes jarmin a tool to minimize application jar files. The tool
takes a class path and main entry point and will produce copies of the jar
files on the classpath that have unreferenced classes removed. The tool is
build on the ASM framework and uses abstract interpretation to handle
reflection and other complex program behaviours. The tool also has a set of
plugins to model complex behaviours like annotation bean loading.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>