Socket connection opened on demand only
Previously, the agent was opening port 9002 by default,
just in case a client would want to connect to it.
Since this is a security vulnerability, the default
behavior has been changed to not open any port.
Should the user want to connect to the agent over the
network, the agent needs to be started with the port
number as an option on the command line.
e.g.: java -agentpath:path=portNo:9002
The client, which previously was using the default port
9002, now needs to specify a port explicitly because
there is no longer a default port number.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>