Implement app callback feature
This commit gives the application the ability to register a
callback function with the agent, such that when an event is
triggered and data collected, that data is sent to the application
through the callback.
To accomodate this change the structure of a command changed
to allow the "callback" keyword, which will define the class,
method name and signature of the callback.
Example:
  {
    "functionality": "monitorEvents",
    "command": "start",
    "callback": {
      "class": "T",
      "method": "callback",
      "signature": "(Ljava/lang/String;)V"
    }
  }
The callback needs to be a static Java method that accepts one
java/lang/String parameter (the data to be sent) and the class
that implements the callback needs to be loaded by the system
classloader so that the class is discoverable through the name
only.
The commit also implements some code restructuring to introduce
a `class EventConfig` which will be used to store (and cache)
information about each type of event that is monitored.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>