add a head field with object hashcode
When an app is profiled over a time period,
a number of sections are produced, potentially
for the same lock. However, there is no key
field in the JSON data that help us to associate
those together.

Add a "monitorHash" field with the hashcode of the object
which is the current subject of contention. According
to the JVMTI spec, the hashcode is guarenteed to be
unique throughout its life span.

Fixes: https://github.com/eclipse/openj9-utils/issues/22

Signed-off-by: Gireesh Punathil <gpunathi@in.ibm.com>