Fix MethodEnter support for Perf Agent
In order to enable method tracing we need to set the
`can_generate_method_entry_events` capability. For some
reason this capability can be enabled only during the
OnLoad phase. Therefore, the agent must be loaded from
the very beginning, and late attach will not do.

This commit also adds the ability to print the stack trace
up to the specified depth.

Examples of commands:
```
  {
    "functionality": "methodEntryEvents",
    "command": "start",
    "delay": 1,
    "stackTraceDepth": 5
  },
  {
    "functionality": "methodEntryEvents",
    "command": "stop",
    "delay": 5
  }
```

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>