Implement JLM support in the perf agent
This commit adds JLM (Java Lock Monitor) support to the
OpenJ9 Perf Agent. The new command must look like this:
{
    "command": "start",
    "delay": 1,
    "functionality": "jlm"
}
At the moment, the only option that could be sent with this
command is the "delay". All other keywords will be ignored.
Typically, the user sends a "start" command for the "jlm"
functionality and, after a while (specified by the delay),
sends a "stop" command. E.g.:
{
    "command": "stop",
    "delay": 10,
    "functionality": "jlm"
}
Once the stop command is received, the agent will ask the JVM
to send a JLM dump which contains information about all the
monitors in the JVM. The agent parses that dump and creates
a more readable output encoded as json. In this output, the
monitors that were not acquired at least once will be skipped.
The monitors are grouped by their type: Java monitors and
Raw monitors (aka system monitors).
Example of output:
```
{
  "body": {
    "JLM size": 24881,
    "javaMonitors": [
      {
        "enterCount": 142780,
        "holdTime": 34360262656,
        "monitorName": "[00007F8E640024A8] Work@00000007FFF77000 (Object)",
        "recursiveCount": 0,
        "slowCount": 1002,
        "spinCount": 51723505,
        "tag": 0,
        "yieldCount": 1560634
      }
    ],
    "rawMonitors": [
      {
        "enterCount": 1,
        "holdTime": 0,
        "monitorName": "[00007F8ED001A598] GCExtensions::_lightweightNonReentrantLockPoolMutex",
        "recursiveCount": 0,
        "slowCount": 0,
        "spinCount": 0,
        "tag": 0,
        "yieldCount": 0
      },
      {
        "enterCount": 4,
        "holdTime": 0,
        "monitorName": "[00007F8ED001B828] JIT-SmallBlockListsMonitor",
        "recursiveCount": 0,
        "slowCount": 0,
        "spinCount": 0,
        "tag": 0,
        "yieldCount": 0
      },
...
      {
        "enterCount": 1,
        "holdTime": 0,
        "monitorName": "[00007F8ED00A1B10] MM_MemorySubSpace:_lock",
        "recursiveCount": 0,
        "slowCount": 0,
        "spinCount": 0,
        "tag": 0,
        "yieldCount": 0
      }
    ]
  },
```
This commit also factors out the code that determines the IBM JVMTI
extensions, so that we call that code only once, reducing overhead.

Closes: #24

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>