Add option to turn on/off method enter/exit capability
Since method enter/exit tracing capability can interfere
with the JIT compiler optimizer, it's better to turn off
this capability by default, but allow the user to turn it
on, if she wants to use it.
This commit adds a new option for the agent called:
`methodEnterCapability` with two possible values "on" or "off".
The default is "off".
Note that enabling method enter/exit tracing capability does
not enable tracing itself. Starting/stopping the tracing
needs to be done with json commands like with any other tracing.

Fixes: #57

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>