Prevent crashes at JVM shutdown
If tracing of an event is enabled while the JVM shuts down,
a crash can happen because the `Server` object is destroyed,
while the tracing thread still wants to use this object.
This commit prevents crashes during shutdown by adding a lock.
When the `Server` object is used, the curent thread will hold
a `serverMonitor` to prevent the thread executing the `VMDeath`
event from destroying the `Server` object. This will also prevent
the output from different tracing threads from being interleaved.

Fixes: #30

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>